services/ai/pyproject.toml
[project]
name = "remy-ai"
version = "0.1.0"
description = "Remy AI orchestration service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.5",
  "uvicorn[standard]>=0.32.1",
  "httpx>=0.27.2",
  "pydantic>=2.9.2",
  "pydantic-settings>=2.6.1",
  "google-genai>=0.3.0",
  "modal>=0.66.0",
  "pillow>=11.0.0",
  "numpy>=2.1.3",
  "python-multipart>=0.0.12",
  "structlog>=24.4.0",
  "orjson>=3.10.11",
  "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
  "pytest-httpx>=0.33.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
  "freezegun>=1.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
